Implement basic type declaration parsing in frontend

+++ backend.hh
@@ -82,7 +82,7 @@ private:
     
     // Helper methods
     void emit(Opcode op, uint32_t lineNumber, 
-              int32_t intValue = 0, 
+              int64_t intValue = 0, 
               float floatValue = 0.0f, 
               bool boolValue = false, 
               const std::string& stringValue = "");
